feat(package): correct implementation of debouncing, modularizing helper functions into utils
